+++ CHANGELOG.md
@@ -25,8 +25,12 @@ changes.
   + Check presence of state token (ST) and that it's consistent against datum.
   + Reduce cost of `commitTx` by using the initial script as input reference.
   + Moved check to reimburse commits to head validator and ensure its completeness.
+<<<<<<< HEAD
   + Remove snapshot number and utxo hash from Close and Contest reedemer as they
     were already present in Closed datum.
+=======
+  + Check no tokens are minted/burnt in v-head for close, contest, commit and collectCom tx.
+>>>>>>> 67e6f2daf (Generate better values)
 
 - **BREAKING** Change the way tx validity and contestation deadline is constructed for close transactions:
   + There is a new hydra-node flag `--contestation-period` expressed in seconds
@@ -63,7 +67,6 @@ changes.
 
 -  `hydra-cluster` executable can be used to provide a local cardano "network" with `--devnet` argument
 
-- Check no tokens are burnt in v-head for close, contest and collectCom tx.
 
 ## [0.8.1] - 2022-11-17